+++ Phoenix/src/ECS/components.h
@@ -119,12 +119,23 @@ namespace Phoenix::Component{
 		Phoenix::UUID save_texture;
 
 		Material() = default;
-		Material(const BasicMaterial& _material)
-			: material(_material) {}
+		Material(const BasicMaterial& mat)
+			: material(mat) {}
 	};
 
 
 	struct Cube : public Material{};
 	struct Plane : public Material{};
+
+
+
+	struct Card{
+		Phoenix::UUID texture;
+		glm::vec2 scale{1.0f, 1.0f};
+		
+		Card() = default;
+		Card(const Phoenix::UUID& uuid) 
+			: texture(uuid) {};
+	};
 	
 }